delete Header.jsx and organize folders
